+++ browser/resources/js/components.js
@@ -1,7 +1,7 @@
 /* eslint-disable */(function (factory) {
 	typeof define === 'function' && define.amd ? define(factory) :
 	factory();
-}(function () { 'use strict';
+}((function () { 'use strict';
 
 	const utility = globalThis.$;
 	utility.cnsl = function(...args) {
@@ -759,4 +759,4 @@
 		pageTitle
 	});
 
-}));
+})));