+++ config/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+	devMode: true,
+	homepage: 'https://social.sentivate.com'
+};