+++ server/chunkMessage/index.js
@@ -1,0 +1,6 @@
+module.exports = (server) => {
+	const {
+		logImprt,
+	} = server;
+	logImprt('PARSE MESSAGE CHUNK', __dirname);
+};