+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 | SENTIVATE - CHANGELOG     |
 | :----: |
 | ![Sentivate Logo](https://raw.githubusercontent.com/sentivate/SentivateAlphaNetwork/master/resources/logo.png)|
-|[Sentivate](https://sentivate.com) - [Twitter](https://twitter.com/sentivate) - [Instagram](https://instagram.com/sntvt) - [Arity](https://arity.company)  - [Arity Twitter](https://twitter.com/TheRealArity)|
+|[Sentivate](https://sentivate.com) - [Twitter](https://twitter.com/sentivate) - [Instagram](https://instagram.com/sntvt) |
 ---
 
 ## Unreleased
@@ -20,21 +20,26 @@
 - Latest Stable with Browser [`6151b50`](https://github.com/sentivate/SentivateAlphaNetwork/commit/6151b501b3461202c86ef6aa063a86906979b951)
 - Network Simulation [`90d69b1`](https://github.com/sentivate/SentivateAlphaNetwork/commit/90d69b14edf0845b718c19fdd3d3e38da66338e6)
 - August Push [`d08b943`](https://github.com/sentivate/SentivateAlphaNetwork/commit/d08b94396817e0c8e00585851d2f51af38664ac8)
+- Packet Structure Update [`221ee9e`](https://github.com/sentivate/SentivateAlphaNetwork/commit/221ee9e79b9d8da22d79e4bac61fe61397f31050)
 - Changelog + Dependencies [`e1eb86f`](https://github.com/sentivate/SentivateAlphaNetwork/commit/e1eb86f1b917bc8a96e9cb0c80d2505837342f5a)
 - Removed n [`0a55554`](https://github.com/sentivate/SentivateAlphaNetwork/commit/0a55554532e4b7c41bbc053cd6efecc2ef224f31)
+- Client Module Cleanup [`7b3e1f3`](https://github.com/sentivate/SentivateAlphaNetwork/commit/7b3e1f3a7daedae4eb16ab7dfa50a29416ae3712)
 - Swap Protocol [`abb3faa`](https://github.com/sentivate/SentivateAlphaNetwork/commit/abb3faa55fb83f8d2bcb4860c2df4bf9f146be75)
 - Module Cleanup [`65123b1`](https://github.com/sentivate/SentivateAlphaNetwork/commit/65123b16453b24d51195a2eefe11d40b51e8ae33)
 - Latest [`5620744`](https://github.com/sentivate/SentivateAlphaNetwork/commit/56207442678fef7404dd3573d2ed8df056be3c6c)
 - Updated & Removed [`d839c5a`](https://github.com/sentivate/SentivateAlphaNetwork/commit/d839c5a249c14103d0d9a9864012f62fc23e83ae)
+- Browser enabled [`d8433ef`](https://github.com/sentivate/SentivateAlphaNetwork/commit/d8433ef7a7b55b64172835c5a0e8ff500db3e309)
 - Update .eslintrc [`0bb17c7`](https://github.com/sentivate/SentivateAlphaNetwork/commit/0bb17c7a9ef3d2d0306a98c849df349972d5a5ef)
 - Removed lzutf8 [`2a96066`](https://github.com/sentivate/SentivateAlphaNetwork/commit/2a96066e11bf7e287b616daf4f096569e9c38004)
 - Create SECURITY.md [`6e75ff8`](https://github.com/sentivate/SentivateAlphaNetwork/commit/6e75ff86cbafaa1e030824a4e77eafd17493b3ca)
 - Test certs [`73978d6`](https://github.com/sentivate/SentivateAlphaNetwork/commit/73978d6d5f8e2cc1c2ce0e25f9bc9c12e6b59e29)
 - Update README.md [`94d1d9c`](https://github.com/sentivate/SentivateAlphaNetwork/commit/94d1d9ca822e32545b39b0f95917b01028a70665)
 - Update README.md [`18f1a50`](https://github.com/sentivate/SentivateAlphaNetwork/commit/18f1a50bfb25066bafe6ddca2dc8b5e144d16031)
+- Sentivate [`d3468b9`](https://github.com/sentivate/SentivateAlphaNetwork/commit/d3468b92a64f414636e8cce1f1545f9bdf81bd92)
 - Organize test scripts [`eb2d402`](https://github.com/sentivate/SentivateAlphaNetwork/commit/eb2d402897e03bb62d666dde5be7fcfb387ce4ba)
 - Readme Update [`95ca894`](https://github.com/sentivate/SentivateAlphaNetwork/commit/95ca89443f0d63a7ab062e5eea87b8059f107063)
 - Update README.md [`18359db`](https://github.com/sentivate/SentivateAlphaNetwork/commit/18359db9242b842df377c91a5f59363f7a8459ea)
+- Update package.json [`17d228b`](https://github.com/sentivate/SentivateAlphaNetwork/commit/17d228b4c2a08d53d3efecca00033e6fa963f5f9)
 - Update .eslintrc [`feb70c7`](https://github.com/sentivate/SentivateAlphaNetwork/commit/feb70c79ccf1ed4438d70c663347aa94e399701d)
 - Update README.md [`b69e94e`](https://github.com/sentivate/SentivateAlphaNetwork/commit/b69e94e3f202efcbe68bf32bcd77d147dec07a2f)
 - Update README.md [`1903e4d`](https://github.com/sentivate/SentivateAlphaNetwork/commit/1903e4d340af2eb93b72c29f919519b85424e8f4)